run-length encoding implemented

+++ value.h
@@ -6,8 +6,8 @@
 typedef double Value;
 
 typedef struct {
-  int capacity;
-  int count;
+  size_t capacity;
+  size_t count;
   Value* values;
 } ValueArray;
